pyproject.toml
[tool.poetry]
name = "oc_p12_epicevents"
version = "0.1.0"
description = ""
authors = ["Clement MARCEL <clmtmarcel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django-admin = "^2.0.1"
Django = "^4.1.5"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
